requirements.txt
biopython
pandas
scikit-bio
matplotlib
seaborn
flask